\name{digits}
\docType{data}
\alias{digits}
\title{Handwritten digits}
\description{Euclidean distances between 1000 handwritten digits.}
\usage{data(digits)}
\format{A \code{\link[stats]{dist}} object with the dissimilarity matrix.}
\references{Hull JJ (1994) A database for handwritten text recognition research. \emph{IEEE Trans Pattern Anal Mach Intell} 16:550--554.}
