\name{protein}
\docType{data}
\alias{protein}
\title{Protein sequences}
\description{Inverse sequence alignment scores from the basic local alignment search tool (BLAST) between 124 protein sequences.}
\usage{data(protein)}
\format{A \code{\link[stats]{dist}} object with the dissimilarity matrix.}
\references{Pollack JD, Li Q, Pearl DK (2005) Taxonomic utility of a phylogenetic analysis of phosphoglycerate kinase proteins of archaea, bacteria, and eukaryota: Insights by bayesian analyses. \emph{Mol Phylogenet Evol} 35:420--430.}
