\name{radar}
\docType{data}
\alias{radar}
\title{Radar signals}
\description{Euclidean distances between 351 radar signals.}
\usage{data(radar)}
\format{A \code{\link[stats]{dist}} object with the dissimilarity matrix.}
\references{Sigillito VG, Wing SP, Hutton LV, Baker KB (1989) Classification of radar returns from the ionosphere using neural networks. \emph{Johns Hopkins Apl Tech Dig} 10:262--266.}
