\name{tpe-package}
\alias{tpe-package}
\alias{tpe}
\docType{package}
\title{Tree preserving embedding}
\description{This package implements the greedy approximation for tree preserving embedding.}
\details{The functions \code{\link{tpe1d}} and \code{\link{tpe2d}} compute one-dimensional and two-dimensional tree preserving embeddings respectively.}
\references{Shieh AD, Hashimoto TB, Airoldi EM (2011) Tree preserving embedding. \emph{Proc Natl Acad Sci USA} 108:16916--16921.}
